Organize imports for geogig plugin
